requirements.txt
streamlit
langchain
langchain-ollama
python-dotenv
pydantic
PyPDF2